Warnings cleanup.

+++ variables/svs_variables.c
@@ -221,8 +221,8 @@ void gcCollectArrays(svsVM *s) {
 
 void printArrays(svsVM *s) {
   puts("Arrays:");
-  for(int32_t x = 1; x <= s->varArrayLen; x++) {
-    printf("%u : %u \n", x, s->varArray[x].val_s);
+  for(uint32_t x = 1; x <= s->varArrayLen; x++) {
+    printf("%u : %ld \n", (unsigned int)x, s->varArray[x].val_s);
   }
   puts("End");
 }
